+++ R/nfsa_validation.R
@@ -27,6 +27,7 @@
 nfsa_validation <- function(time_min,
                             time_max = lubridate::ymd(lubridate::today()),
                             output_sel = here::here("output", "logs")){
+library(dplyr)
 
   cli::cli_progress_message("Searching for files in the server...")
   validated <- list.files("I:/econ/NFSA/VALID/results/",